[RCP-374] :recycle: 로딩 애니메이션 수정작업 진행
