+++ app/src/main/java/com/recipia/aos/ui/components/common/AnimatedPreloader.kt
@@ -1,0 +1,40 @@
+package com.recipia.aos.ui.components.common
+
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.animateLottieCompositionAsState
+import com.airbnb.lottie.compose.rememberLottieComposition
+import com.recipia.aos.R
+
+/**
+ * 커스텀 로딩 인디케이터 (로티 사용)
+ */
+@Composable
+fun AnimatedPreloader(modifier: Modifier = Modifier) {
+
+    val preloaderLottieComposition by rememberLottieComposition(
+        LottieCompositionSpec.RawRes(
+            R.raw.loading // 여기에 애니메이션 리소스를 지정합니다.
+        )
+    )
+
+    val preloaderProgress by animateLottieCompositionAsState(
+        preloaderLottieComposition,
+        iterations = LottieConstants.IterateForever,
+        isPlaying = true
+    )
+
+    // Lottie 애니메이션을 화면에 표시
+    LottieAnimation(
+        composition = preloaderLottieComposition,
+        progress = preloaderProgress,
+        modifier = modifier.size(150.dp)
+    )
+
+}
