+++ app/src/main/java/com/recipia/aos/ui/components/mypage/MyPageRecipeListScreen.kt
@@ -57,6 +57,7 @@ import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
 import com.recipia.aos.R
 import com.recipia.aos.ui.components.HorizontalDivider
+import com.recipia.aos.ui.components.common.AnimatedPreloader
 import com.recipia.aos.ui.dto.RecipeListResponseDto
 import com.recipia.aos.ui.model.mypage.MyPageViewModel
 import com.recipia.aos.ui.model.recipe.bookmark.BookMarkViewModel
@@ -160,7 +161,7 @@ fun MyPageRecipeListScreen(
             ) {
                 // 로딩 중이라면 로딩 인디케이터 표시
                 if (isLoading) {
-                    MyPageAnimatedPreloader(modifier = Modifier.size(100.dp)) // 로딩 바의 크기 조절 가능
+                    AnimatedPreloader(modifier = Modifier.size(100.dp))
                 } else {
                     LazyColumn(
                         contentPadding = PaddingValues(
@@ -193,7 +194,7 @@ fun MyPageRecipeListScreen(
 
 
 @Composable
-fun     MyPageRecipeListItem(
+fun MyPageRecipeListItem(
     item: RecipeListResponseDto,
     bookmarkViewModel: BookMarkViewModel,
     navController: NavController
@@ -333,27 +334,3 @@ fun     MyPageRecipeListItem(
         )
     }
 }
-
-@Composable
-fun MyPageAnimatedPreloader(modifier: Modifier = Modifier) {
-    val preloaderLottieComposition by rememberLottieComposition(
-        LottieCompositionSpec.RawRes(
-            R.raw.loading // 여기에 애니메이션 리소스를 지정합니다.
-        )
-    )
-
-    val preloaderProgress by animateLottieCompositionAsState(
-        preloaderLottieComposition,
-        iterations = LottieConstants.IterateForever,
-        isPlaying = true
-    )
-
-    // Lottie 애니메이션을 화면에 표시합니다.
-    // `modifier` 매개변수를 사용하여 사이즈 조절
-    LottieAnimation(
-        composition = preloaderLottieComposition,
-        progress = preloaderProgress,
-        modifier = modifier.size(100.dp) // 여기에서 원하는 크기로 조절합니다.
-    )
-}
-
